backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Banco de dados
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Excel e processamento
pandas==2.1.3
openpyxl==3.1.2

# Utilitários
python-dateutil==2.8.2
